:recycle: Refactored nested files
