+++ src/Screens/Auth/index.js
@@ -1,0 +1,3 @@
+import AuthScreen from './AuthScreen'
+
+export default AuthScreen
