+++ src/Screens/PlaylistItem/index.js
@@ -1,0 +1,3 @@
+import PlaylistItemScreen from './PlaylistItemScreen'
+
+export default PlaylistItemScreen
