+++ src/Screens/PlaylistList/SwipeablePlaylistItem.js
@@ -2,9 +2,9 @@ import React, { useState } from 'react'
 import { Alert } from 'react-native'
 
 import images from 'res/images'
-import AnimatedSwipeRow from 'Gruvee/src/Screens/Common/AnimatedSwipeRow'
-import SwipeAction from 'Gruvee/src/Screens/Common/SwipeAction'
-import PlaylistItem from '../PlaylistItem/PlaylistItem'
+import AnimatedSwipeRow from 'lib/Common/AnimatedSwipeRow'
+import SwipeAction from 'lib/Common/SwipeAction'
+import PlaylistItem from 'Gruvee/src/screens/PlaylistItem/'
 import * as StyleConstants from 'lib/Helpers/StyleConstants'
 
 const SwipeablePlaylistItem = ({
