+++ src/Screens/PlaylistList/index.js
@@ -1,0 +1,3 @@
+import PlaylistListScreen from './PlaylistListScreen'
+
+export default PlaylistListScreen
