+++ src/Screens/SongItem/index.js
@@ -1,0 +1,3 @@
+import SongItemScreen from './SongItemScreen'
+
+export default SongItemScreen
