+++ src/Screens/SongList/index.js
@@ -1,0 +1,3 @@
+import SongListScreen from './SongListScreen'
+
+export default SongListScreen
