+++ src/lib/Common/AnimatedSwipeRow/AnimatedSwipeRow.js
