+++ src/lib/Common/AnimatedSwipeRow/AnimatedSwipeRow.styles.js
