+++ src/lib/Common/InputModal/InputModal.styles.js
