+++ src/lib/Common/SwipeAction/SwipeAction.js
