+++ src/lib/Common/SwipeAction/SwipeAction.styles.js
