+++ src/lib/Common/SwipeAction/index.js
